requirements.txt
openai
ipykernel
dotenv
azure-ai-ml
azure-common
azure-ai-documentintelligence